Split core client tests into modular includes
