+++ crates/core/src/client/tests/mod.rs
@@ -1,0 +1,9 @@
+include!("client_event.rs");
+include!("builder_compress.rs");
+include!("builder_metadata.rs");
+include!("fallback_invocation.rs");
+include!("fallback_controls.rs");
+include!("client_transfer.rs");
+include!("module_list_request.rs");
+include!("module_list_proxy.rs");
+include!("module_list_auth.rs");
